% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeriveClus.R
\name{DeriveClus}
\alias{DeriveClus}
\title{Title}
\usage{
DeriveClus(casesM, ctrlsM, optimal, DeriveMed, ClusteringMed, save_res = TRUE)
}
\arguments{
\item{casesM}{aaa}

\item{ctrlsM}{bbb}

\item{optimal}{ccc}

\item{DeriveMed}{ddd}

\item{ClusteringMed}{eee}

\item{save_res}{fff}
}
\value{

}
\description{
Title
}
\examples{
library(multtest)
data(golub)
testcases1 <- DeriveClus(t(golub[1:500,1:27]),t(golub[1:500,28:38]),9,DeriveMed="Cases",ClusteringMed="Pam")
}
