% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMCfunc.R
\name{DMCmod}
\alias{DMCmod}
\title{Title}
\usage{
DMCmod(
  rho,
  G,
  N,
  M = 500,
  permnum = 500,
  alpha = 0.05,
  propred = c(0.1, 0.3, 0.5, 0.7, 0.9),
  redpt,
  CVstru,
  adjpower = 1,
  type = "signed",
  save_res = TRUE
)
}
\arguments{
\item{rho}{Cov parameter}

\item{G}{# of Genes}

\item{N}{# of Subjects}

\item{M}{# of Modules}

\item{permnum}{# of Permutations}

\item{alpha}{# Alpha level}

\item{propred}{aa}

\item{redpt}{bb}

\item{CVstru}{cc}

\item{adjpower}{dd}

\item{type}{ee}

\item{save_res}{gg}
}
\value{

}
\description{
Title
}
\examples{
trial1 <- DMCmod(rho=0.7, G=10, N=50, M=10, permnum=10, alpha=0.05, propred=c(0.1,0.3,0.5,0.7,0.9),redpt=c(0,0.3,0.5),CVstru="AR",adjpower=1,type = "signed", save_res = FALSE)
}
