% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testClus.R
\name{testClus}
\alias{testClus}
\title{Title}
\usage{
testClus(
  Derived_Modules,
  permnum,
  compare_med = "adjvscorr",
  adjpower = 1,
  type = "signed",
  save_res = TRUE
)
}
\arguments{
\item{Derived_Modules}{aaaa}

\item{permnum}{bbbbb}

\item{compare_med}{ccccc}

\item{adjpower}{ddddd}

\item{type}{eeeee}

\item{save_res}{fffff}
}
\value{

}
\description{
Title
}
\examples{
library(multtest)
data(golub)
testcases1 <- DeriveClus(t(golub[1:100,1:27]),t(golub[1:100,28:38]),7,DeriveMed="Cases",ClusteringMed="Pam")
trial222 <- testClus(testcases1,100,compare_med="adjvstom",type = "signed",adjpower = 0.5)


}
